api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "chromadb>=1.3.5",
    "fastapi>=0.124.0",
    "google-genai>=1.53.0",
    "pydantic-settings>=2.11.0",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.9",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

